pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "people"
version = "0.1.0-dev"
description = "The Python SDK for the Solid Project"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Pacific" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.27",
    "rdflib>=7.0",
    "PyJWT[crypto]>=2.8",
    "cryptography>=42.0",
]

[project.optional-dependencies]
pandas = ["pandas>=2.0"]
science = ["networkx>=3.0"]
all = ["people[pandas,science]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.0",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/Pacific-Systems-Ltd/people"
Issues = "https://github.com/Pacific-Systems-Ltd/people/issues"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]
ignore = [
    "N807",   # dunder function names in model decorator (intentional)
    "SIM105", # contextlib.suppress vs try/except/pass (readability preference)
    "SIM108", # ternary vs if/else (readability preference)
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
